server/runpod/requirements_runpod.txt
runpod
torch
torchvision
transformers
pillow
safetensors
